Add independent Indices module.
Currently it can only output health.  However, it is the beginning of
an internal refactoring where the subcommands can be passed json
output and not have to call out to ES.  This is extremely helpful if
you only have the json output from a cluster at a previous point in
time.